docker/FASTAPI/requirements.txt
fastapi
uvicorn[standard]
pandas
gunicorn
pydantec
scikit-surprise